int -> long
Changing type of the non-default constructor of "JDKRandomGenerator" in
order to match the signature of "java.util.Random".

+++ src/main/java/org/apache/commons/math4/random/JDKRandomGenerator.java
@@ -42,7 +42,7 @@ public class JDKRandomGenerator extends Random implements RandomGenerator {
      * @param seed initial seed
      * @since 3.6
      */
-    public JDKRandomGenerator(int seed) {
+    public JDKRandomGenerator(long seed) {
         setSeed(seed);
     }
